requirements.txt
asn1crypto==0.24.0
autopep8==1.4.4
cffi==1.12.3
cryptography==2.7
Django==2.2.2
django-cors-headers==3.0.2
django-filter==2.1.0
django-rest-knox==4.1.0
djangorestframework==3.9.4
gunicorn==19.9.0
Markdown==3.1.1
Pillow==6.1.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pycparser==2.19
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.3
